1.0.0 - uploaded code directly from the past project
The code was directly uploaded from a project that was completed four years ago. It needs to be rewritten due to the numerous changes that have occurred over this lengthy period. The old database has been removed and the data must be recollected. Furthermore, it's likely that the structure of the original job posting website has also been updated